restructured files so that files that were components of a screen are now under the Component folder
