+++ components/Background.js
